on edit.blade remember the enctype="" for the data. On the mail/welcomemail the __construict accepts the public User $user, public Post $post.
On the Postcontroller.
once the post is created, assign it to a $post var.

On the update part of the controller.
add the image to validation.
Lastly to change the image, or keep as it is if not changed.
Use the Storage::disk to delete and then Storage:: ->put to recreate